+++ R/project-dag-write.R
@@ -12,14 +12,14 @@ populate_project_dag_write <- function(batch = FALSE, verbose = FALSE) {
     # nocov end
   }
 
-  credential  <- retrieve_credential_testing(2545L, "admin")
+  credential  <- retrieve_credential_testing("dag-write", username = "admin")
 
   project <- REDCapR::redcap_project$new(
     redcap_uri    = credential$redcap_uri,
     token         = credential$token
   )
   path_in_dag <- system.file(
-    "test-data/project-dag/data.csv",
+    "test-data/projects/dag/data.csv",
     package = "REDCapR"
   )
 
@@ -46,23 +46,23 @@ populate_project_dag_write <- function(batch = FALSE, verbose = FALSE) {
 
   # Import the data into the REDCap project
   # testthat::expect_message(
-    returned_object <- if (batch) {
-      REDCapR::redcap_write(
-        ds_to_write                 = ds_to_write,
-        redcap_uri                  = project$redcap_uri,
-        token                       = project$token,
-        verbose                     = verbose,
-        convert_logical_to_integer  = TRUE
-      )
-    } else {
-      REDCapR::redcap_write_oneshot(
-        ds                          = ds_to_write,
-        redcap_uri                  = project$redcap_uri,
-        token                       = project$token,
-        verbose                     = verbose,
-        convert_logical_to_integer  = TRUE
-      )
-    }
+  returned_object <- if (batch) {
+    REDCapR::redcap_write(
+      ds_to_write                 = ds_to_write,
+      redcap_uri                  = project$redcap_uri,
+      token                       = project$token,
+      verbose                     = verbose,
+      convert_logical_to_integer  = TRUE
+    )
+  } else {
+    REDCapR::redcap_write_oneshot(
+      ds                          = ds_to_write,
+      redcap_uri                  = project$redcap_uri,
+      token                       = project$token,
+      verbose                     = verbose,
+      convert_logical_to_integer  = TRUE
+    )
+  }
   # )
 
   # # If uploading the data was successful, then upload the image files.
@@ -94,8 +94,9 @@ clear_project_dag_write <- function(verbose = FALSE) {
     )
     # nocov end
   }
-  path_delete_test_record <-
-    "https://bbmc.ouhsc.edu/redcap/plugins/redcapr/delete_redcapr_dag_write.php"
+
+  path_delete_test_record <- retrieve_plugins("delete_dag")
+  # "https://redcap-dev-2.ouhsc.edu/redcap/plugins/redcapr/delete_redcapr_dag_write.php"
 
   # Returns a boolean value if successful
   was_successful <- !httr::http_error(path_delete_test_record)