+++ R/project-delete-multiple-arm.R
@@ -10,14 +10,15 @@ populate_project_delete_multiple_arm <- function(verbose = FALSE) {
     # nocov end
   }
 
-  credential  <- retrieve_credential_testing(2627L)
+  credential  <- retrieve_credential_testing("arm-multiple-delete")
 
   project <- REDCapR::redcap_project$new(
     redcap_uri    = credential$redcap_uri,
     token         = credential$token
   )
   path_in <- system.file(
-    "test-data/delete-multiple-arm/data.csv",
+    "test-data/projects/arm-multiple-delete/data.csv",
+    # "test-data/delete-multiple-arm/data.csv",
     package = "REDCapR"
   )
 
@@ -63,8 +64,8 @@ clear_project_delete_multiple_arm <- function(verbose = TRUE) {
     )
     # nocov end
   }
-  path_delete_test_record <-
-    "https://bbmc.ouhsc.edu/redcap/plugins/redcapr/delete_redcapr_delete_multiple_arm.php"
+  path_delete_test_record <- retrieve_plugins("delete_arm_multiple")
+  # "https://redcap-dev-2.ouhsc.edu/redcap/plugins/redcapr/delete_redcapr_delete_multiple_arm.php"
 
   # Returns a boolean value if successful
   was_successful <- !httr::http_error(path_delete_test_record)