+++ tests/testthat/test-read-batch-longitudinal.R
@@ -1,8 +1,12 @@
 library(testthat)
 
-credential          <- retrieve_credential_testing(212L)
+credential          <- retrieve_credential_testing("longitudinal")
 project             <- redcap_project$new(redcap_uri=credential$redcap_uri, token=credential$token)
 update_expectation  <- FALSE
+
+if (credential$redcap_uri != "https://redcap-dev-2.ouhsc.edu/redcap/api/") {
+  testthat::skip("Skipping tests with lots of consecutive new lines on non-dev server")
+}
 
 test_that("smoke", {
   testthat::skip_on_cran()
@@ -30,7 +34,7 @@ test_that("smoke", {
 })
 
 test_that("so-example-data-frame-retrieval", {
-  path_expected <- "test-data/project-longitudinal/expected/so-example-data-frame-retrieval.R"
+  path_expected <- "test-data/projects/longitudinal/expected/so-example-data-frame-retrieval.R"
 
   actual <- tibble::tibble(a=1:5, b=6:10)
 
@@ -43,7 +47,7 @@ test_that("so-example-data-frame-retrieval", {
 
 test_that("default", {
   testthat::skip_on_cran()
-  path_expected <- "test-data/project-longitudinal/expected/default.R"
+  path_expected <- "test-data/projects/longitudinal/expected/default.R"
   expected_outcome_message <- "\\d+ records and \\d+ columns were read from REDCap in \\d+(\\.\\d+\\W|\\W)seconds\\."
 
   ###########################
@@ -89,7 +93,7 @@ test_that("default", {
 
 test_that("filter-numeric", {
   testthat::skip_on_cran()
-  path_expected <- "test-data/project-longitudinal/expected/filter-numeric.R"
+  path_expected <- "test-data/projects/longitudinal/expected/filter-numeric.R"
   expected_outcome_message <- "\\d+ records and \\d+ columns were read from REDCap in \\d+(\\.\\d+\\W|\\W)seconds\\."
 
   filter <- "[bmi] > 25"
@@ -140,7 +144,7 @@ test_that("filter-numeric", {
 
 test_that("filter-character", {
   testthat::skip_on_cran()
-  path_expected <- "test-data/project-longitudinal/expected/filter-character.R"
+  path_expected <- "test-data/projects/longitudinal/expected/filter-character.R"
   expected_outcome_message <- "\\d+ records and \\d+ columns were read from REDCap in \\d+(\\.\\d+\\W|\\W)seconds\\."
 
   filter <- "[email] = 'zlehnox@gmail.com'"