+++ tests/testthat/test-read-batch-survey.R
@@ -1,6 +1,6 @@
 library(testthat)
 
-credential          <- retrieve_credential_testing(817L)
+credential          <- retrieve_credential_testing("survey")
 project             <- redcap_project$new(redcap_uri=credential$redcap_uri, token=credential$token)
 update_expectation  <- FALSE
 
@@ -36,7 +36,7 @@ test_that("smoke", {
 })
 
 test_that("so-example-data-frame-retrieval", {
-  path_expected <- "test-data/project-survey/expected/so-example-data-frame-retrieval.R"
+  path_expected <- "test-data/projects/survey/expected/so-example-data-frame-retrieval.R"
 
   actual <- tibble::tibble(a=1:5, b=6:10) # saveRDS(actual, file.path("./inst", directory_relative, file_name))
 
@@ -49,7 +49,7 @@ test_that("so-example-data-frame-retrieval", {
 
 test_that("default", {
   testthat::skip_on_cran()
-  path_expected <- "test-data/project-survey/expected/default.R"
+  path_expected <- "test-data/projects/survey/expected/default.R"
   expected_outcome_message <- "\\d+ records and \\d+ columns were read from REDCap in \\d+(\\.\\d+\\W|\\W)seconds\\."
 
   ###########################
@@ -62,11 +62,18 @@ test_that("default", {
       verbose               = FALSE
     )
 
-  if (update_expectation) save_expected(returned_object1$data, path_expected)
+  expect_true(all(!is.na(returned_object1$data$prescreening_survey_timestamp)))
+  expect_s3_class(returned_object1$data$prescreening_survey_timestamp, "POSIXct")
+  d1 <-
+    returned_object1$data |>
+    dplyr::select(
+      -prescreening_survey_timestamp
+    )
+
+  if (update_expectation) save_expected(d1, path_expected)
   expected_data_frame <- retrieve_expected(path_expected)
 
-  expect_equal(returned_object1$data, expected=expected_data_frame, label="The returned data.frame should be correct", ignore_attr = TRUE) # dput(returned_object1$data)
-  expect_true(all(!is.na(returned_object1$data$prescreening_survey_timestamp)))
+  expect_equal(d1, expected=expected_data_frame, label="The returned data.frame should be correct", ignore_attr = TRUE) # dput(returned_object1$data)
   expect_true(returned_object1$success)
   expect_match(returned_object1$status_codes, regexp="200", perl=TRUE)
   expect_true(returned_object1$records_collapsed=="", "A subset of records was not requested.")
@@ -86,8 +93,15 @@ test_that("default", {
       verbose               = FALSE
     )
 
-  expect_equal(returned_object2$data, expected=expected_data_frame, label="The returned data.frame should be correct", ignore_attr = TRUE) # dput(returned_object2$data)
-  expect_true(all(!is.na(returned_object1$data$prescreening_survey_timestamp)))
+  expect_true(all(!is.na(returned_object2$data$prescreening_survey_timestamp)))
+  expect_s3_class(returned_object2$data$prescreening_survey_timestamp, "POSIXct")
+  d2 <-
+    returned_object2$data |>
+    dplyr::select(
+      -prescreening_survey_timestamp
+    )
+
+  expect_equal(d2, expected=expected_data_frame, label="The returned data.frame should be correct", ignore_attr = TRUE) # dput(returned_object2$data)
   expect_true(returned_object2$success)
   expect_match(returned_object2$status_codes, regexp="200", perl=TRUE)
   expect_true(returned_object2$records_collapsed=="", "A subset of records was not requested.")