+++ tests/testthat/test-read-oneshot-eav.R
@@ -3,6 +3,10 @@ library(testthat)
 credential  <- retrieve_credential_testing()
 update_expectation  <- FALSE
 
+if (credential$redcap_uri != "https://redcap-dev-2.ouhsc.edu/redcap/api/") {
+  testthat::skip("Skipping EAV test on non-dev server")
+}
+
 test_that("smoke test", {
   testthat::skip_on_cran()
   suppressMessages({
@@ -94,6 +98,7 @@ test_that("raw", {
   expect_s3_class(returned_object$data, "tbl")
 })
 test_that("raw-and-dag", {
+  testthat::skip("Temporarily turning off DAG on experimental function")
   testthat::skip_on_cran()
   path_expected <- "test-data/specific-redcapr/read-oneshot-eav/raw-and-dag.R"
   expected_outcome_message <- "\\d+ records and \\d+ columns were read from REDCap in \\d+(\\.\\d+\\W|\\W)seconds\\."
@@ -121,6 +126,7 @@ test_that("raw-and-dag", {
   expect_s3_class(returned_object$data, "tbl")
 })
 test_that("label-and-dag", {
+  testthat::skip("Temporarily turning off DAG on experimental function")
   testthat::skip_on_cran()
   path_expected <- "test-data/specific-redcapr/read-oneshot-eav/label-and-dag.R"
   expected_outcome_message <- "\\d+ records and \\d+ columns were read from REDCap in \\d+(\\.\\d+\\W|\\W)seconds\\."
@@ -132,7 +138,7 @@ test_that("label-and-dag", {
       raw_or_label                = "label",
       export_data_access_groups   = TRUE,
       verbose                     = FALSE
-  )
+    )
 
   if (update_expectation) save_expected(returned_object$data, path_expected)
   expected_data_frame <- retrieve_expected(path_expected)
@@ -233,7 +239,7 @@ test_that("filter-character", {
 })
 test_that("blank-for-gray-status-true", {
   testthat::skip_on_cran()
-  credential_blank_for_gray  <- retrieve_credential_testing(3003L)
+  credential_blank_for_gray  <- retrieve_credential_testing("blank-for-gray-status")
   path_expected <- "test-data/specific-redcapr/read-oneshot-eav/blank-for-gray-true.R"
   expected_outcome_message <- "\\d+ records and \\d+ columns were read from REDCap in \\d+(\\.\\d+\\W|\\W)seconds\\."
 
@@ -245,10 +251,16 @@ test_that("blank-for-gray-status-true", {
       verbose                     = FALSE
     )
 
-  if (update_expectation) save_expected(returned_object$data, path_expected)
+  d <-
+    returned_object$data |>
+    dplyr::select(
+      -mugshot # Don't compare file IDs across servers.
+    )
+
+  if (update_expectation) save_expected(d, path_expected)
   expected_data_frame <- retrieve_expected(path_expected)
 
-  expect_equal(returned_object$data, expected=expected_data_frame, label="The returned data.frame should be correct", ignore_attr = TRUE) # dput(returned_object$data)
+  expect_equal(d, expected=expected_data_frame, label="The returned data.frame should be correct", ignore_attr = TRUE) # dput(returned_object$data)
   expect_equal(returned_object$status_code, expected=200L)
   expect_equal(returned_object$raw_text, expected="", ignore_attr = TRUE) # dput(returned_object$raw_text)
   expect_true(returned_object$records_collapsed=="", "A subset of records was not requested.")
@@ -260,7 +272,7 @@ test_that("blank-for-gray-status-true", {
 })
 test_that("blank-for-gray-status-false", {
   testthat::skip_on_cran()
-  credential_blank_for_gray  <- retrieve_credential_testing(3003L)
+  credential_blank_for_gray  <- retrieve_credential_testing("blank-for-gray-status")
   path_expected <- "test-data/specific-redcapr/read-oneshot-eav/blank-for-gray-false.R"
   expected_outcome_message <- "\\d+ records and \\d+ columns were read from REDCap in \\d+(\\.\\d+\\W|\\W)seconds\\."
 
@@ -272,10 +284,16 @@ test_that("blank-for-gray-status-false", {
       verbose                     = FALSE
     )
 
-  if (update_expectation) save_expected(returned_object$data, path_expected)
+  d <-
+    returned_object$data |>
+    dplyr::select(
+      -mugshot # Don't compare file IDs across servers.
+    )
+
+  if (update_expectation) save_expected(d, path_expected)
   expected_data_frame <- retrieve_expected(path_expected)
 
-  expect_equal(returned_object$data, expected=expected_data_frame, label="The returned data.frame should be correct", ignore_attr = TRUE) # dput(returned_object$data)
+  expect_equal(d, expected=expected_data_frame, label="The returned data.frame should be correct", ignore_attr = TRUE) # dput(returned_object$data)
   expect_equal(returned_object$status_code, expected=200L)
   expect_equal(returned_object$raw_text, expected="", ignore_attr = TRUE) # dput(returned_object$raw_text)
   expect_true(returned_object$records_collapsed=="", "A subset of records was not requested.")
@@ -285,7 +303,6 @@ test_that("blank-for-gray-status-false", {
   expect_true(returned_object$success)
   expect_s3_class(returned_object$data, "tbl")
 })
-
 test_that("date-range", {
   testthat::skip_on_cran()
   path_expected <- "test-data/specific-redcapr/read-oneshot-eav/default.R"